extract interface from table printer
