+++ shvatka/infrastructure/printer/table.py
@@ -1,35 +1,20 @@
 import typing
-from dataclasses import dataclass
-from datetime import datetime, time, timedelta
+from io import BytesIO
 from typing import Any
 
 from openpyxl import Workbook
 from openpyxl.worksheet.worksheet import Worksheet
 
-DATETIME_EXCEL_FORMAT = "HH:MM:SS"
+from shvatka.core.interfaces.printer import Table, TablePrinter
 
 
-@dataclass(kw_only=True, frozen=True)
-class CellAddress:
-    column: int
-    row: int
+class ExcellPrinter(TablePrinter):
 
-    def shift(self, rows: int = 0, columns: int = 0) -> "CellAddress":
-        return CellAddress(column=self.column + columns, row=self.row + rows)
-
-    def to_dict(self) -> dict[str, int]:
-        return {"row": self.row, "column": self.column}
-
-
-@dataclass(kw_only=True)
-class Cell:
-    value: str | datetime | int | time
-    format: str | None = None
-
-
-@dataclass(kw_only=True)
-class Table:
-    fields: dict[CellAddress, Cell]
+    def print_table(self, table: Table) -> BytesIO:
+        result = BytesIO()
+        print_table(table, result)
+        result.seek(0)
+        return result
 
 
 def print_table(table: Table, file: typing.Any) -> None:
@@ -48,10 +33,3 @@ def resize_columns(worksheet: Worksheet):
     for col in worksheet.columns:  # type: Any  # =(
         new_len = max([2, *[len(str(cell.value or "")) for cell in col]])
         worksheet.column_dimensions[col[0].column_letter].width = new_len
-
-
-def as_time(td: timedelta) -> time:
-    hours = td.seconds // 3600
-    minutes = (td.seconds - hours * 3600) // 60
-    seconds = td.seconds - hours * 3600 - minutes * 60
-    return time(hours, minutes, seconds, td.microseconds)
