+++ shvatka/tgbot/views/results/level_times.py
@@ -7,13 +7,9 @@ from shvatka.core.models import dto
 from shvatka.core.utils.datetime_utils import trim_tz
 from shvatka.core.utils.exceptions import GameNotFinished
 from shvatka.infrastructure.printer.table import (
-    CellAddress,
-    Cell,
-    Table,
     print_table,
-    DATETIME_EXCEL_FORMAT,
-    as_time,
 )
+from shvatka.core.interfaces.printer import DATETIME_EXCEL_FORMAT, CellAddress, Cell, Table, as_time
 
 FIRST_TEAM_NAME = CellAddress(row=3, column=1)
 GAME_NAME = CellAddress(row=1, column=1)
